test: add shared classic battle mocks
