+++ tests/helpers/classicBattle/commonMocks.js
@@ -1,0 +1,22 @@
+import { vi } from "vitest";
+
+vi.mock("../../../src/helpers/motionUtils.js", () => ({
+  shouldReduceMotionSync: () => true
+}));
+
+vi.mock("../../../src/utils/scheduler.js", () => ({
+  onFrame: (cb) => {
+    const id = setTimeout(() => cb(performance.now()), 16);
+    return id;
+  },
+  onSecondTick: (cb) => {
+    const id = setInterval(() => cb(performance.now()), 1000);
+    return id;
+  },
+  cancel: vi.fn((id) => {
+    clearTimeout(id);
+    clearInterval(id);
+  }),
+  start: vi.fn(),
+  stop: vi.fn()
+}));
