+++ tests/helpers/classicBattle/countdownReset.test.js
@@ -1,26 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
+import "./commonMocks.js";
 import { createBattleHeader, createBattleCardContainers } from "../../utils/testUtils.js";
 import { createRoundMessage, createSnackbarContainer, createTimerNodes } from "./domUtils.js";
-vi.mock("../../../src/utils/scheduler.js", () => ({
-  onFrame: (cb) => {
-    const id = setTimeout(() => cb(performance.now()), 16);
-    return id;
-  },
-  onSecondTick: (cb) => {
-    const id = setInterval(() => cb(performance.now()), 1000);
-    return id;
-  },
-  cancel: (id) => {
-    clearTimeout(id);
-    clearInterval(id);
-  },
-  start: vi.fn(),
-  stop: vi.fn()
-}));
-
-vi.mock("../../../src/helpers/motionUtils.js", () => ({
-  shouldReduceMotionSync: () => true
-}));
 
 vi.mock("../../../src/helpers/classicBattle/uiHelpers.js", async () => {
   const actual = await vi.importActual("../../../src/helpers/classicBattle/uiHelpers.js");
