+++ tests/helpers/classicBattle/debugPanel.test.js
@@ -1,4 +1,5 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
+import "./commonMocks.js";
 
 vi.mock("../../../src/helpers/classicBattle/opponentController.js", () => ({
   getOpponentCardData: vi.fn()
@@ -30,8 +31,6 @@ vi.mock("../../../src/helpers/setupScoreboard.js", () => ({
   updateTimer: vi.fn()
 }));
 vi.mock("../../../src/helpers/battle/index.js", () => ({ showResult: vi.fn() }));
-vi.mock("../../../src/helpers/motionUtils.js", () => ({ shouldReduceMotionSync: () => true }));
-vi.mock("../../../src/utils/scheduler.js", () => ({ onFrame: vi.fn(), cancel: vi.fn() }));
 vi.mock("../../../src/helpers/classicBattle/selectionHandler.js", () => ({
   handleStatSelection: vi.fn()
 }));
