+++ tests/helpers/classicBattle/opponentDelay.test.js
@@ -1,8 +1,6 @@
 // @vitest-environment jsdom
 import { describe, it, expect, vi, beforeEach } from "vitest";
-vi.mock("../../../src/helpers/motionUtils.js", () => ({
-  shouldReduceMotionSync: () => true
-}));
+import "./commonMocks.js";
 
 let showMessage;
 let clearMessage;
