+++ tests/helpers/classicBattle/selectionPrompt.test.js
@@ -1,26 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
+import "./commonMocks.js";
 import { setupClassicBattleDom } from "./utils.js";
 import { applyMockSetup } from "./mockSetup.js";
-
-vi.mock("../../../src/helpers/motionUtils.js", () => ({
-  shouldReduceMotionSync: () => true
-}));
-vi.mock("../../../src/utils/scheduler.js", () => ({
-  onFrame: (cb) => {
-    const id = setTimeout(() => cb(performance.now()), 16);
-    return id;
-  },
-  onSecondTick: (cb) => {
-    const id = setInterval(() => cb(performance.now()), 1000);
-    return id;
-  },
-  cancel: (id) => {
-    clearTimeout(id);
-    clearInterval(id);
-  },
-  start: vi.fn(),
-  stop: vi.fn()
-}));
 
 let timerSpy;
 let fetchJsonMock;
